+++ examples/jukebox.py
@@ -1,20 +1,4 @@
 #!/usr/bin/env python
-#
-# $Id$
-#
-# Copyright 2009 Doug Winter
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
 
 import cmd
 import readline
@@ -78,7 +62,7 @@ class JukeboxUI(cmd.Cmd, threading.Thread):
                 playlist = self.jukebox.starred
             for i, t in enumerate(playlist):
                 if t.is_loaded():
-                    print "%3d %s" % (i, t.name())
+                    print "%3d %s - %s" % (i, t.artists()[0].name(), t.name())
                 else:
                     print "%3d %s" % (i, "loading...")