+++ spotify/__init__.py
@@ -1,3 +1,4 @@
+__version__ = '1.3'
 
 class SpotifyError(Exception):
     pass