+++ spotify/manager/playlist.py
@@ -1,17 +1,3 @@
-# Copyright 2011 Antoine Pierlot-Garcin
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 class SpotifyPlaylistManager:
     """
     Handles Spotify playlists callbacks. To implement you own callbacks,
@@ -28,11 +14,11 @@ class SpotifyPlaylistManager:
         playlist.add_tracks_added_callback(self.tracks_added, self, userdata)
         playlist.add_tracks_moved_callback(self.tracks_moved, self, userdata)
         playlist.add_tracks_removed_callback(self.tracks_removed, self,
-                                                                    userdata)
+            userdata)
 
     def unwatch(self, playlist, userdata=None):
         """
-        Stop listenning to events on the playlist.
+        Stop listening to events on the playlist.
         """
         try:
             playlist.remove_callback(self.tracks_added, userdata)
@@ -41,20 +27,20 @@ class SpotifyPlaylistManager:
         except:
             pass
 
-######## CALLBACKS ########
+### Callbacks
 
     def tracks_added(self, playlist, tracks, position, userdata):
         """
         Callback
 
+        Called when tracks are inserted in the playlist.
+
         :param playlist:    playlist on which the event occured
         :type playlist:     :class:`spotify.Playlist`
         :param tracks:      list of added tracks
-        :type tracks:       :class:`spotify.Track`
+        :type tracks:       list of :class:`spotify.Track`
         :param position:    position in which the tracks where inserted
         :type position:     :class:`int`
-
-        Called when tracks are inserted in the playlist.
         """
         pass
 
@@ -62,14 +48,14 @@ class SpotifyPlaylistManager:
         """
         Callback
 
+        Called when tracks are moved in the playlist.
+
         :param playlist:    playlist on which the event occured
         :type playlist:     :class:`spotify.Playlist`
         :param tracks:      list of moved tracks
-        :type tracks:       :class:`spotify.Track`
+        :type tracks:       list of :class:`spotify.Track`
         :param new_position: new position of the tracks
-        :type new_position:     :class:`int`
-
-        Called when tracks are moved in the playlist.
+        :type new_position: :class:`int`
         """
         pass
 
@@ -77,12 +63,12 @@ class SpotifyPlaylistManager:
         """
         Callback
 
+        Called when tracks are removed from the playlist.
+
         :param playlist:    playlist on which the event occured
         :type playlist:     :class:`spotify.Playlist`
         :param tracks:      list of removed tracks
-        :type tracks:       :class:`spotify.Track`
-
-        Called when tracks are removed from the playlist.
+        :type tracks:       list of :class:`spotify.Track`
         """
         pass
 
@@ -90,11 +76,11 @@ class SpotifyPlaylistManager:
         """
         Callback
 
-        :param playlist:    playlist on which the event occured
-        :type playlist:     :class:`spotify.Playlist`
-
         Called when a playlist has been renamed.
         :meth:`spotify.Playlist.name()` can be used to find out the new name.
+
+        :param playlist:    playlist on which the event occured
+        :type playlist:     :class:`spotify.Playlist`
         """
         pass
 
@@ -102,17 +88,17 @@ class SpotifyPlaylistManager:
         """
         Callback
 
-        :param playlist:    playlist on which the event occured
-        :type playlist:     :class:`spotify.Playlist`
-
         Called when state changed for a playlist.
 
         There are three states that trigger this callback:
 
-            - Collaboration for this playlist has been turned on or off
-            - The playlist started having pending changes, or all pending
-                changes have now been committed
-            - The playlist started loading, or finished loading
+        - Collaboration for this playlist has been turned on or off.
+        - The playlist started having pending changes, or all pending changes
+          have now been committed.
+        - The playlist started loading, or finished loading.
+
+        :param playlist:    playlist on which the event occured
+        :type playlist:     :class:`spotify.Playlist`
         """
         pass
 
@@ -120,16 +106,16 @@ class SpotifyPlaylistManager:
         """
         Callback
 
-        :param playlist:    playlist on which the event occured
-        :type playlist:     :class:`spotify.Playlist`
-        :param done:        wether the update is finished
-        :type done:         :class:`bool`
-
         Called when a playlist is updating or is done updating.
 
         This is called before and after a series of changes are applied to the
         playlist. It allows e.g. the user interface to defer updating until the
         entire operation is complete.
+
+        :param playlist:    playlist on which the event occured
+        :type playlist:     :class:`spotify.Playlist`
+        :param done:        whether the update is finished
+        :type done:         :class:`bool`
         """
         pass
 
@@ -137,17 +123,19 @@ class SpotifyPlaylistManager:
         """
         Callback
 
-        :param playlist:    playlist on which the event occured
-        :type playlist:     :class:`spotify.Playlist`
-
         Called when metadata for one or more tracks in a playlist has been
         updated.
+
+        :param playlist:    playlist on which the event occured
+        :type playlist:     :class:`spotify.Playlist`
         """
         pass
 
     def track_created_changed(self, playlist, position, user, when, userdata):
         """
         Callback
+
+        Called when create time and/or creator for a playlist entry changes.
 
         :param playlist:    playlist on which the event occured
         :type playlist:     :class:`spotify.Playlist`
@@ -157,8 +145,6 @@ class SpotifyPlaylistManager:
         :type user:         :class:`spotify.User`
         :param when:        time in seconds since the UNIX Epoch
         :type when:         :class:`int`
-
-        Called when create time and/or creator for a playlist entry changes.
         """
         pass
 
@@ -166,14 +152,14 @@ class SpotifyPlaylistManager:
         """
         Callback
 
+        Called when message attribute for a playlist entry changes.
+
         :param playlist:    playlist on which the event occured
         :type playlist:     :class:`spotify.Playlist`
         :param position:    track's position in the playlist
         :type position:     :class:`int`
         :param message:     new message
         :type message:      :class:`unicode`
-
-        Called when message attribute for a playlist entry changes.
         """
         pass
 
@@ -181,14 +167,14 @@ class SpotifyPlaylistManager:
         """
         Callback
 
+        Called when seen attribute for a playlist entry changes.
+
         :param playlist:    playlist on which the event occured
         :type playlist:     :class:`spotify.Playlist`
         :param position:    track's position in the playlist
         :type position:     :class:`int`
         :param seen:        new seen attribute
         :type seen:         :class:`bool`
-
-        Called when seen attribute for a playlist entry changes.
         """
         pass
 
@@ -196,12 +182,12 @@ class SpotifyPlaylistManager:
         """
         Callback
 
+        Called when playlist description has changed.
+
         :param playlist:    playlist on which the event occured
         :type playlist:     :class:`spotify.Playlist`
         :param description: new description
         :type description:  :class:`unicode`
-
-        Called when playlist description has changed.
         """
         pass
 
@@ -209,10 +195,10 @@ class SpotifyPlaylistManager:
         """
         Callback
 
+        Called when playlist subscribers changes (count or list of names).
+
         :param playlist:    playlist on which the event occured
         :type playlist:     :class:`spotify.Playlist`
-
-        Called when playlist subscribers changes (count or list of names).
         """
         pass
 
@@ -220,11 +206,11 @@ class SpotifyPlaylistManager:
         """
         Callback
 
+        Called when playlist image has changed.
+
         :param playlist:    playlist on which the event occured
         :type playlist:     :class:`spotify.Playlist`
         :param image:       image id of the new image
         :type image:        :class:`str`
-
-        Called when playlist image has changed.
         """
         pass