+++ src/artist.c
@@ -1,20 +1,3 @@
-/* $Id$
- *
- * Copyright 2009 Doug Winter
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at 
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
 #include <Python.h>
 #include <structmember.h>
 #include "libspotify/api.h"
@@ -65,19 +48,19 @@ Artist_name(Artist * self)
     const char *s = sp_artist_name(self->_artist);
 
     if (!s)
-        return NULL;
-    return PyString_FromString(s);
+        Py_RETURN_NONE;
+    return PyUnicode_FromString(s);
 }
 
-PyObject *
+static PyObject *
 Artist_str(PyObject *self)
 {
     Artist *a = (Artist *) self;
     const char *s = sp_artist_name(a->_artist);
 
     if (!s)
-        return NULL;
-    return PyString_FromString(s);
+        Py_RETURN_NONE;
+    return PyUnicode_FromString(s);
 }
 
 static PyMethodDef Artist_methods[] = {