+++ src/playlist.c
@@ -1,20 +1,3 @@
-/* $Id$
- *
- * Copyright 2009 Doug Winter
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
 #include <Python.h>
 #include <structmember.h>
 #include <libspotify/api.h>
@@ -55,6 +38,7 @@ Playlist_FromSpotify(sp_playlist * spl)
         (Playlist *) PyObject_CallObject((PyObject *)&PlaylistType, NULL);
     playlist->_playlist = spl;
     sp_playlist_add_ref(spl);
+    sp_playlist_set_autolink_tracks(spl, 1);
     return (PyObject *)playlist;
 }
 
@@ -732,7 +716,7 @@ Playlist_name(Playlist * self)
 {
     const char *name = sp_playlist_name(self->_playlist);
 
-    return Py_BuildValue("s", name);
+    return PyUnicode_FromString(name);
 }
 
 static PyObject *