+++ src/search.c
@@ -1,20 +1,3 @@
-/* $Id$
- *
- * Copyright 2009 Doug Winter
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
 #include <Python.h>
 #include <structmember.h>
 #include "libspotify/api.h"
@@ -65,7 +48,7 @@ Results_is_loaded(Results * self)
 static PyObject *
 Results_did_you_mean(Results * self)
 {
-    return PyString_FromString(sp_search_did_you_mean(self->_search));
+    return PyUnicode_FromString(sp_search_did_you_mean(self->_search));
 }
 
 static PyObject *
@@ -132,13 +115,13 @@ Results_total_tracks(Results * self)
 static PyObject *
 Results_query(Results * self)
 {
-    return PyString_FromString(sp_search_query(self->_search));
+    return PyUnicode_FromString(sp_search_query(self->_search));
 }
 
 PyObject *
 Results_str(PyObject *self)
 {
-    return PyString_FromString(sp_search_query(((Results *) self)->_search));
+    return PyUnicode_FromString(sp_search_query(((Results *) self)->_search));
 }
 
 static PyMethodDef Results_methods[] = {