+++ src/track.c
@@ -1,20 +1,3 @@
-/* $Id$
- *
- * Copyright 2009 Doug Winter
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
 #include <Python.h>
 #include <structmember.h>
 #include "libspotify/api.h"
@@ -62,7 +45,7 @@ Track_str(PyObject *oself)
     Track *self = (Track *) oself;
     const char *s = sp_track_name(self->_track);
 
-    return Py_BuildValue("s", s);
+    return PyUnicode_FromString(s);
 }
 
 static PyObject *
@@ -106,7 +89,7 @@ Track_name(Track * self)
 {
     const char *s = sp_track_name(self->_track);
 
-    return Py_BuildValue("s", s);
+    return PyUnicode_FromString(s);
 }
 
 static PyObject *