+++ tests/test_albumbrowser.py
@@ -1,22 +1,8 @@
-#
-# Copyright 2011 Antoine Pierlot-Garcin
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 import unittest
 from spotify._mockspotify import mock_albumbrowse, mock_album, mock_artist
 from spotify._mockspotify import mock_session
 from spotify import Album
+from tests import SkipTest
 
 callback_called = False
 callback_userdata = None
@@ -46,8 +32,8 @@ class TestAlbumbrowser(unittest.TestCase):
         assert browser[1].name() == 'bar'
         assert browser[2].name() == 'baz'
 
-    @unittest.skip('Not implemented')
     def test_callback(self):
+        raise SkipTest
         callback_called = False
         browser = mock_albumbrowse(self.session, self.album, 0, self.callback,
                                    self)