+++ tests/test_artist.py
@@ -1,18 +1,4 @@
-# $Id$
-#
-# Copyright 2009 Doug Winter
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at 
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+# encoding: utf-8
 
 import unittest
 from spotify._mockspotify import Artist, mock_artist
@@ -34,3 +20,6 @@ class TestArtist(unittest.TestCase):
         artist = mock_artist("test_name", 1)
         self.assertEqual(artist.name(), "test_name")
 
+    def test_unicode(self):
+        artist = mock_artist(u'æâ€êþÿ', 1)
+        self.assertEqual(artist.name(), u'æâ€êþÿ')