+++ tests/test_playlist.py
@@ -1,18 +1,4 @@
-# $Id$
-#
-# Copyright 2009 Doug Winter
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+# encoding: utf-8
 
 import unittest
 from spotify import _mockspotify
@@ -52,6 +38,10 @@ class TestPlaylist(unittest.TestCase):
     def test_name(self):
         p1 = _mockspotify.mock_playlist("foo", [])
         self.assertEqual(p1.name(), "foo")
+
+    def test_unicode(self):
+        p1 = _mockspotify.mock_playlist(u'æâ€êþÿ', [])
+        self.assertEqual(p1.name(), u'æâ€êþÿ')
 
     def test_len(self):
         p1 = self._mock_track("foo")