+++ docs/changes.rst
@@ -2,6 +2,20 @@
 Changes
 =======
 
+v1.3 (2011-06-11)
+=================
+
+It has only been four days since the v1.2 release, but we would like to get the
+change from bytestrings to unicode objects released before more projects start
+using pyspotify, as this change is really backwards incompatible.
+
+- All strings returned by pyspotify has been changed from UTF-8 encoded
+  bytestrings to unicode objects.
+- Track autolinking enabled for all playlists.
+- Add :attr:`spotify.__version__` which exposes the current pyspotify version.
+  The API version of the libspotify used is already available as
+  :attr:`spotify.api_version`.
+
 v1.2 (2011-06-07)
 =================